refactor: close page on every request and handle errors
The scraper often gets a page crashed error which must be captued as an event rather than try/catch.
In this commit, I add an event listener that throws the error so it can be captured in try/catch.
Due to stability issues of the scraper, the cache was removed for the time being